restructured project and separate routes and controllers
