--- server/src/job/job.ts
@@ -1,8 +1,0 @@
-export default interface Job {
-  title: string;
-  company: string;
-  location: string;
-  date: string;
-  employmentType: string;
-  link: string;
-}
