+++ server/src/services/scraperService.ts
@@ -1,0 +1,11 @@
+import { scrapHipsters } from '../scrapers/hipsters';
+import { scrapStackoverflow } from '../scrapers/stackOverflow';
+import { scrapGeekhunterData } from '../scrapers/geekhunter';
+import { scrapProgramathor } from '../scrapers/programathor';
+
+export {
+    scrapGeekhunterData,
+    scrapHipsters,
+    scrapProgramathor,
+    scrapStackoverflow,
+};
