--- server/src/services/stackOverflow.ts
@@ -1,145 +1,0 @@
-import cheerio from 'cheerio';
-import fetchHtml from './request';
-import { formatToBRdate } from '../utils/dateFormater';
-import Job from '../job/job';
-import Dictionary from '../types/dictionary';
-
-const scrapData = async () => {
-  const $ = await getPageSelector();
-
-  const firstPageJobs = extractJobs($, getUrl().baseUrl);
-
-  const jobsPerPage = firstPageJobs.length;
-
-  const pagesCount = getPagesCount($, jobsPerPage);
-
-  const otherPagesJobs = await getAllJobs(2, pagesCount);
-
-  const jobs = [...firstPageJobs, ...otherPagesJobs];
-
-  return jobs;
-};
-
-const getAllJobs = async (pageStart = 1, pagesCount: number) => {
-  let jobs = Array<Job>();
-  const { baseUrl } = getUrl();
-
-  for (let page = pageStart; page <= pagesCount; page++) {
-    const pageSelector = await getPageSelector(page);
-    const pageJobs = extractJobs(pageSelector, baseUrl);
-
-    jobs = [...jobs, ...pageJobs];
-  }
-
-  return jobs;
-};
-
-const getPageSelector = async (page = 1): Promise<CheerioStatic> => {
-  const { url } = getUrl(page);
-  const html = await fetchHtml(url);
-
-  const $ = cheerio.load(html);
-
-  return $;
-};
-
-const getPagesCount = ($: CheerioStatic, amountPerPage: number): number => {
-  const jobsAmountText = $('div#index-hed > div.js-search-title')
-    .find('span.description')
-    .first()
-    .text();
-
-  const total = parseInt(jobsAmountText.match(/\d+/)[0]);
-  const pageCount = Math.ceil(total / amountPerPage);
-
-  return pageCount;
-};
-
-const extractJobs = ($: CheerioStatic, baseUrl: string): Job[] => {
-  const jobs = Array<Job>();
-
-  const results = $('body')
-    .find("div[class='listResults']")
-    .find('div.-job > div.grid > div.fl1 ');
-
-  results.each((index, element) => {
-    const title = $(element)
-      .find('a.stretched-link')
-      .text()
-      .trim();
-
-    const company = $(element)
-      .find('h3.fc-black-700')
-      .children()
-      .first()
-      .text()
-      .trim();
-
-    const location = $(element)
-      .find('h3.fc-black-700 > span.fc-black-500')
-      .text()
-      .trim();
-
-    const jobPath = $(element)
-      .find('a.stretched-link')
-      .attr('href')
-      .trim();
-
-    const link = `${baseUrl}${jobPath}`;
-
-    const createdAt = $(element)
-      .find('div.mt2')
-      .children()
-      .first()
-      .text()
-      .trim();
-
-    const date = solveDate(createdAt);
-
-    const job: Job = {
-      title,
-      company,
-      location,
-      date,
-      employmentType: '',
-      link,
-    };
-
-    jobs.push(job);
-  });
-
-  return jobs;
-};
-
-const solveDate = (createdAt: string): string => {
-  const match = createdAt.match(/(\d+).*([a-z])\s/);
-  const key = match ? match[2] : createdAt;
-  const time = parseInt(match?.[1]) || 0;
-
-  const daysToSub = timeSwitch(key, time);
-  const date = new Date();
-
-  date.setDate(date.getDate() - daysToSub);
-
-  return formatToBRdate(date);
-};
-
-const timeSwitch = (key: string, time: number): number => {
-  const options: Dictionary = {
-    h: 0,
-    d: time,
-    yesterday: 1,
-  };
-
-  return options[key] || 0;
-};
-
-const getUrl = (page = 1) => {
-  const baseUrl = 'https://stackoverflow.com/';
-  return {
-    baseUrl,
-    url: `${baseUrl}/jobs?r=true&sort=p&pg=${page}`,
-  };
-};
-
-export { scrapData as scrapStackoverflow };
