+++ server/src/shared/fetchHtml.ts
@@ -1,0 +1,13 @@
+import axios from 'axios';
+
+const fetchHTML = async (url: string) => {
+    try {
+        const { data } = await axios.get(url);
+        return data;
+    } catch (error) {
+        console.log('error on REQUEST');
+        console.log(error);
+    }
+};
+
+export default fetchHTML;
