+++ server/src/shared/types/dictionary.ts
@@ -1,3 +1,3 @@
 export default interface Dictionary {
-  [key: string]: number;
+    [key: string]: number;
 }
