+++ server/src/shared/types/job.ts
@@ -1,0 +1,8 @@
+export default interface Job {
+    title: string;
+    company: string;
+    location: string;
+    date: string;
+    employmentType: string;
+    link: string;
+}
