+++ server/src/utils/dateFormater.ts
@@ -1,7 +1,7 @@
 import Intl from 'intl';
 
 const formatToBRdate = (date: Date): string => {
-  return new Intl.DateTimeFormat('pt-BR').format(date);
+    return new Intl.DateTimeFormat('pt-BR').format(date);
 };
 
 export { formatToBRdate };
